fix(ConferenceFixture): let the tests fail when there is no hangup button
If we catch Throwable and then quit the driver there will be no screenshot
taken for the test which has failed. It was not the right thing to ignore
the hangup button being not clickable. If there is a specific case where
this could be ignored then it should be handled accordingly instead.